Changed error messages:  now don't print anything below main() unless the new option --show-below-main is on.  It's on by default.  It also affects suppressions generated with --gen-suppressions=yes.  Updated reg tests accordingly.
Also updated docs for this.  And added some missing command-line args to docs.
Also compartmentalised the options a little in the docs, and rearranged the
order of options in the usage message, in anticipation of a bigger
rearrangement that will be necessary soon -- to distinguish options used by all
skins from those used by error-checking skins, to skin-specific ones.


git-svn-id: svn://svn.valgrind.org/valgrind/trunk@2020 a5019735-40e9-0310-863c-91ae7b9d1cf9