More fixes for the cancellation wrappers in libpthread - if looking for the original function with RTLD_NEXT doesn't work then try looking for the __libc_ version of the function the RTLD_DEFAULT instead.
The reason for this is that, contrary to the dlsym documentation, it
seems that RTLD_NEXT doesn't always seem to find the definition that
would have been used if it weren't for the override. This is particularly
common wihen libpthread is pulled in implicitly by a dependency from
another library.

This should hopefully fix bug #85658.


git-svn-id: svn://svn.valgrind.org/valgrind/trunk@2595 a5019735-40e9-0310-863c-91ae7b9d1cf9