Fix problems with very long timeouts given when waiting on a mutex or condition variable. The pthread routines now use a timeout of 0xfffffffe if the user asks for something longer than that otherwise we will wrap around and actually get a much shorter timeout.
The scheduler has also been changed so that it it now limits itself to
a timeout of 0x7fffffff when working how how long to poll for. This won't
affect how long a thread actually sleeps for as we'll just wind up waiting
a bit more on the next pass round the loop.

This fixes bug 76845.


git-svn-id: svn://svn.valgrind.org/valgrind/trunk@2781 a5019735-40e9-0310-863c-91ae7b9d1cf9