This patch by Philippe Waroquiers, philippe.waroquiers@skynet.be replaces r12124. His analysis of the testcase failure:
I think I understand what is happening: even if the ptrace invoker functionality
is not needed, the timeout to invoke might expire, which then leads
to a message produced by vgdb if ptrace is restricted by the kernel.

I think the best way to fix this is to add the option --max-invoke-ms=0 to vgdb. 


git-svn-id: svn://svn.valgrind.org/valgrind/trunk@12130 a5019735-40e9-0310-863c-91ae7b9d1cf9