Add the lwpid in the scheduler status information E.g. we now have: Thread 1: status = VgTs_Runnable (lwpid 15782) ==15782==    at 0x8048EB5: main (sleepers.c:188) client stack range: [0xBE836000 0xBE839FFF] client SP: 0xBE838F80 valgrind stack top usage: 10264 of 1048576
Thread 2: status = VgTs_WaitSys (lwpid 15828)
==15782==    at 0x2E9451: ??? (syscall-template.S:82)
==15782==    by 0x8048AD3: sleeper_or_burner (sleepers.c:84)
==15782==    by 0x39B924: start_thread (pthread_create.c:297)
==15782==    by 0x2F107D: clone (clone.S:130)
client stack range: [0x442F000 0x4E2EFFF] client SP: 0x4E2E338
valgrind stack top usage: 2288 of 1048576

This allows to attach with GDB to the good lwpid in case
you want to examine the valgrind state rather than the guest state.

(it is needed to attach to the specific lwpid as valgrind is not
linked with lib pthread, so GDB cannot discover the threads
of the process).



git-svn-id: svn://svn.valgrind.org/valgrind/trunk@15231 a5019735-40e9-0310-863c-91ae7b9d1cf9